added more commands and changed scan_ip.py name
